requirements.txt
transformers>=4.40.0
datasets>=2.19.0
tokenizers>=0.19.0
evaluate>=0.4.0
accelerate>=0.29.0
torch>=2.2.0
gradio>=4.26.0
python-dotenv>=1.0.0
huggingface_hub>=0.22.0
sentencepiece>=0.2.0
Pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.4.0
